Move ProxyFromEnv to main.
